pyproject.toml
[project]
name = "thin-supervisor"
version = "0.1.0"
description = "Thin tmux sidecar supervisor for long-running AI coding agent workflows"
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
keywords = ["ai", "agents", "supervisor", "workflow", "tmux"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
llm = ["litellm>=1.40"]
dev = ["pytest>=7.0"]

[project.scripts]
thin-supervisor = "supervisor.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["supervisor"]
